Priorities done

+++ website/result.py
@@ -5,6 +5,7 @@ from .models import Event,Date,Result
 from . import db
 from ortools.linear_solver import pywraplp
 from datetime import datetime
+from sqlalchemy.orm import sessionmaker
 
 res = Blueprint('result' , __name__)
 
@@ -12,19 +13,21 @@ keys = list()
 values = list()
 
 count_id =0
-@res.route('/result' , methods = ['GET','POST'])
+@res.route('/result' , methods = ['GET' , 'POST'])
 def result():
 
+    id = current_user.id
+    if request.method =='POST':
+         
+        result = list()
         weight = {}
-
         global count_id
-        id = current_user.id
-        try:
+        
+        try:    
             Result.query.filter_by(user_id=id).delete()
             db.session.commit()
         except:
             db.session.rollback()
-
         handle_request()
 
         for i in range(len(keys)) :
@@ -37,8 +40,7 @@ def result():
         eventList= Event.query.all() 
         DateList = Date.query.all()
 
-
-        #the array with the selected events 
+        # the array with the selected events 
         selected_array=list()
 
         # print(sorted_weight.keys())
@@ -71,7 +73,6 @@ def result():
                     if date_key not in date_vars:
                         date_vars[date_key] = solver.BoolVar('')
             
-
         # Add the constraint that each event can only be scheduled once
         for event_key in event_vars.keys():
             solver.Add(solver.Sum([event_vars[event_key][date_key] for date_key in event_vars[event_key].keys()]) <= 1)
@@ -79,20 +80,26 @@ def result():
         # constraint that each date is different in the result
         for date_key in date_vars.keys():
             solver.Add(solver.Sum([event_vars[event_key][date_key] for event_key in event_vars.keys() if date_key in event_vars[event_key]]) <= 1)
-            
+             
+        values_prior=list()
+        for p in sorted_weight.values():
+            values_prior.append(p)
+                
         # Create the objective function
-        c=0
         objective = solver.Objective()
+        c=0
         for event_key in event_vars.keys():
-            for date_key in event_vars[event_key].keys():
-                priority = sorted_weight[keys[c]] # get the priority weight for this event
-                objective.SetCoefficient(event_vars[event_key][date_key], priority) # set the coefficient for this event and date combination
+            
+            priority = values_prior[c] # get the priority weight for this event
+            print(values_prior[c])
             c+=1
+            for date_key in event_vars[event_key].keys():
+                objective.SetCoefficient(event_vars[event_key][date_key], priority) # set the coefficient for this event and date combination
+            
         objective.SetMaximization()
 
         # Solve the problem
         p = 0
-        data_result = list()
         status = solver.Solve()
         if status == solver.OPTIMAL:
             # Print the solution
@@ -100,37 +107,57 @@ def result():
                 for date_key in event_vars[event_key].keys():
                     if event_vars[event_key][date_key].solution_value() == 1:
                         p += 1
-                        print(str(p) + "." + event_key + " on " + date_key)
+                        evid= Event.query.filter_by(name=event_key).first()
+                        eventID=evid.id
+                        obj  =  Date.query.filter_by(day=date_key).filter_by(event_id=eventID).first()
+                        time = obj.time
+                        result_string =str(p)+". "+ event_key + " on " + date_key +" @ " + time
+                        # print( result_string)
+                    
+                        result.append(result_string)
+                        
+                        r = Result.query.filter_by(result = result_string).first()
+                        if not r:
+                            new_result = Result(result = result_string , user_id =id )
+                            try:
+                                db.session.add(new_result)
+                            except:
+                                db.session.rollback()
+                                print("Raised Exeption result!!")
+                                raise
         else:
             print("Solver returned non-optimal solution")
+        db.session.commit()
 
-            
-        return  render_template('/result.html', user=  current_user)#, selected = final_selection)
-    
+        # print(result)
+        
+        return  render_template('/result.html', user=  current_user, result= result)
+    else:
+        final_result = list()
+        result = Result.query.filter_by(user_id = id).all()
+        for result_item in result:
+            final_result.append(result_item.result+'\n')
+        return  render_template('/result.html', user=  current_user, result = final_result)
+
+#Get from index.js Map     
 def handle_request():
+    
     r = list()
 
     keys.clear()
     values.clear()
     
     counter= 0 
-    if request.method == 'POST':
-        data = request.get_json()
-        map_data = json.dumps(data['map'])
+    
+    data = request.get_json()
+    map_data = json.dumps(data['map'])
 
-        r.append([int (s) for s  in re.findall(r'\d+',map_data)])
-        for i in r :
-                for x in i:
-                        if counter % 2==0:
-                                keys.append(x)  
-                        else :
-                                values.append(x)
-                        counter+=1
-
-def priority_weight(priority):
-    if priority == 1:
-        return 3
-    elif priority == 2:
-        return 2
-    else:
-        return 1  
+    r.append([int (s) for s  in re.findall(r'\d+',map_data)])
+    for i in r :
+            for x in i:
+                    if counter % 2==0:
+                            keys.append(x)  
+                    else :
+                            values.append(x)
+                    counter+=1
+
